Move manual event handling to camera state / tracked object
